# new/normaliser/openapi_orchestrate.yaml
openapi: 3.0.3
info:
  title: CRM Prompt Normaliser
  version: 1.0.0
  description: 'Converts a free-form CRM question into a validated execution plan for The Wave Group CRM
    tools. Returns the metric, the tool that owns it, resolved start and end dates, canonical entity filters,
    and the exact query text each backend tool parses correctly.

    This tool MUST be called first on every CRM question, before any lead, opportunity, event, task, case
    or targets-vs-actuals tool. It performs no database access and returns in milliseconds.'
servers:
- url: https://cuspal-retroserrate-nancey.ngrok-free.dev
  description: Normaliser service
paths:
  /normalise:
    post:
      operationId: normalise_crm_query
      summary: Normalise a CRM question into a tool execution plan
      description: 'Call this FIRST for every CRM data question, before selecting or calling any other
        tool.


        Send the user''s question exactly as they typed it. The response tells you which tool to call,
        with which query text, over which resolved dates.


        If "ok" is false, do not call any tool. Rephrase "clarification" naturally for the user.


        If "ok" is true, execute every entry in "calls". Pass each "canonical_text" to its tool as the
        question parameter WITHOUT ANY MODIFICATION. Do not reword, reorder, re-punctuate or shorten it.
        The wording is chosen to match what that specific backend parser accepts, and altering it causes
        the tool to return data for the wrong period without raising an error.


        Never show the user anything from "diagnostics"; it is internal engineering output. Put the period
        in the table heading using "period_display", and add no caveat, warning or note about it.'
      tags:
      - normaliser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NormaliseRequest'
      responses:
        '200':
          description: A validated execution plan, or a clarification request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NormaliseResponse'
        '422':
          description: The request body was malformed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    description: What was wrong with the request.
  /health:
    get:
      operationId: normaliser_health
      summary: Check the normaliser is running
      description: Liveness check. Reports whether the canonical entity vocabulary is loaded. Not needed
        during normal conversation.
      tags:
      - normaliser
      responses:
        '200':
          description: Service status.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: healthy when the service is up.
                  vocabulary_loaded:
                    type: boolean
                    description: False means entity filters will not be resolved and vocab_build.py needs
                      to be run.
                  facets:
                    type: array
                    description: Entity categories available for filtering.
                    items:
                      type: string
  /format_funnel:
    post:
      operationId: format_funnel_tables
      summary: Render a funnel response as its two display tables
      description: 'Call this after ANY funnel tool returns, from within the funnel agent that just called
        it.


        A funnel tool returns ONE flat record per row holding the stage counts and the conversion ratios
        interleaved in alphabetical order, so TL:VL and MB:MD sit among Total Leads and Junk Leads. It
        has to be shown as TWO tables: Funnel Metrics, then Funnel Conversion Ratios. This endpoint does
        that split deterministically.


        Send the funnel tool''s response in "response" EXACTLY as it came back -- all seventeen keys,
        unreordered, nothing dropped. Send the tool you called in "tool" and the period in "heading".
        Return the resulting "markdown", "metrics_table" and "ratios_table" to the master unchanged.


        It applies the fixed column order, exactly five ratio columns (TL:VL, VL:SOL, SOL:MB, MB:MD, MD:SD),
        Indian digit grouping, and a Total row copied from the response''s own totals block rather than
        summed. A single-row funnel gets no S.No and no Total row.


        If "missing_ratio_columns" comes back non-empty, the payload lost keys in transit -- resend the
        untouched response rather than displaying a short table.'
      tags:
      - normaliser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FunnelFormatRequest'
      responses:
        '200':
          description: Both tables, ready to display.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FunnelFormatResponse'
        '422':
          description: The request body was malformed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    description: What was wrong with the request.
components:
  schemas:
    NormaliseRequest:
      type: object
      required:
      - query
      properties:
        query:
          type: string
          description: The user's CRM question, exactly as they typed it. For a follow-up such as "what
            about Veridia?", first rewrite it into a standalone question carrying the previous metric
            and period, then send that.
          example: Total leads for wave city between 1 April 2026 and 30 June 2026
        today:
          type: string
          description: Reference date as YYYY-MM-DD, used to resolve relative periods such as "last quarter".
            Omit in normal use so the server date applies.
          example: '2026-08-26'
        decompose:
          type: boolean
          default: true
          description: Split questions naming several entities or periods into separate tool calls. Leave
            true; this is what removes the need for the user to type "separately".
    ToolCall:
      type: object
      description: One backend tool invocation to perform.
      properties:
        tool:
          type: string
          description: 'The backend tool to invoke. CRM-Data tools: lead_report, opportunity_report, event_report,
            task_report, case_report, targetvsactuals. CRM-Funnel tools: lead_funnel, project_funnel,
            product_funnel, source_funnel, subsource_funnel, lead_user_funnel, sales_user_funnel.'
        agent:
          type: string
          description: The collaborator agent that owns this tool, either CRM-Data or CRM-Funnel. Delegate
            the call to that agent.
        metric:
          type: string
          description: Canonical metric key, for example total_leads or sales_done.
        metric_label:
          type: string
          description: Human-readable metric name for use in headings.
        canonical_text:
          type: string
          description: Pass this to the tool as its question parameter, byte for byte. Never edit it.
            For example "April, May and June 2026" must keep the word "and", and "fy 2025" must not be
            expanded into a date range, because the backend parsers silently resolve the altered forms
            to the wrong period.
        start_date:
          type: string
          description: Resolved period start, YYYY-MM-DD. Authoritative. Pass it as a parameter where
            the tool accepts one.
        end_date:
          type: string
          description: Resolved period end, YYYY-MM-DD.
        period_kind:
          type: string
          description: range, month_list, quarter_list, year_list or unresolved.
        period_label:
          type: string
          description: Internal label. Do not show this to the user.
        comparison:
          type: string
          description: none, month_on_month, quarter_on_quarter or year_on_year.
        groupings:
          type: array
          description: Breakdown dimensions requested, for example project, product, source, owner.
          items:
            type: string
        filters:
          type: object
          description: Canonical entity values to filter on, keyed by category such as project or product.
            Rows outside these values must not be shown to the user.
          additionalProperties:
            type: array
            items:
              type: string
        period_display:
          type: string
          description: Clean period text for the table heading, for example 'FY2025-26' or 'April 2026
            to June 2026'. Use it in the heading when the returned rows match it. When the backend returns
            a different span, build the heading from the actual rows instead. Never add a caveat about
            the difference.
    NormaliseResponse:
      type: object
      properties:
        ok:
          type: boolean
          description: True means a plan was produced and you should execute "calls". False means call
            nothing and show "clarification" to the user.
        raw:
          type: string
          description: The question as received.
        call_count:
          type: integer
          description: Number of tool calls in the plan.
        calls:
          type: array
          description: The tool calls to execute, in order. Execute every one and wait for all results
            before answering.
          items:
            $ref: '#/components/schemas/ToolCall'
        agents:
          type: array
          description: Collaborator agents these calls belong to.
          items:
            type: string
        clarification:
          type: string
          description: Present when ok is false. Rephrase it naturally for the user and call no tool.
            This is the only field whose content may reach them.
        decomposed:
          type: boolean
          description: True when the question was split into several calls because it named multiple entities
            or periods.
        diagnostics:
          type: array
          items:
            type: string
          description: INTERNAL ONLY. Engineering notes about how the plan was built. Never display these,
            never quote them and never paraphrase them into a warning. The period in period_display already
            tells the user what they need.
    FunnelFormatRequest:
      type: object
      required:
      - response
      properties:
        response:
          type: object
          description: 'The funnel tool''s response, pasted through unchanged. Every shape the seven funnel
            services return is accepted: the single record under lead_funnel, a breakdown keyed by scope
            under product_wise_metrics, and the list forms under source_wise_metrics, sub_source_wise_metrics,
            lead_wise_user_metrics and sales_wise_user_metrics. Do not trim, flatten or reformat it first.'
        tool:
          type: string
          description: 'The plan''s tool value, copied exactly: lead_funnel, product_funnel, project_funnel,
            source_funnel, subsource_funnel, lead_user_funnel or sales_user_funnel. It names the breakdown
            column. Always pass it: project_funnel returns its rows under a key reading product_wise_metrics,
            so without the tool name a project breakdown comes out headed Product.'
          example: lead_funnel
        heading:
          type: string
          description: Period for the table headings, taken from period_display, for example FY2025-26.
            Optional but recommended.
          example: FY2025-26
        show:
          type: string
          default: both
          enum:
          - both
          - metrics
          - ratios
          description: Leave as both, which is right for almost every funnel question. Use ratios only
            when the user's own words said funnel ratios or conversion ratios, and metrics only when they
            said funnel metrics or stage counts. Never narrow on inference.
    FunnelFormatResponse:
      type: object
      properties:
        ok:
          type: boolean
          description: False means no funnel rows were found in the payload. Show the figures another
            way rather than an empty section.
        markdown:
          type: string
          description: BOTH tables, ready to display. Copy this verbatim into the response, the same way
            a graph url is copied. This is the field you print.
        metrics_table:
          type: string
          description: The Funnel Metrics table alone, if you need it separately.
        ratios_table:
          type: string
          description: The Funnel Conversion Ratios table alone, if you need it separately.
        row_count:
          type: integer
          description: Rows rendered. Check your table against it.
        scope_column:
          type: string
          description: The breakdown column used, such as Product, Source, Sub-Source, Project or User.
            Empty for an overall funnel.
        error:
          type: string
          description: Present only when ok is false.
        missing_ratio_columns:
          type: array
          items:
            type: string
          description: Ratio columns the service reports but that were absent from the payload sent here.
            Non-empty means the response was trimmed in transit.
        warning:
          type: string
          description: Present when missing_ratio_columns is non-empty.
tags:
- name: normaliser
  description: Query normalisation for The Wave Group CRM agents.